Ensure we clean up callbacks
When registering a callback, make sure we're also registering EndOfEvent. This ensures we always know when to clean up the callback, but don't clean it up for other callbacks which may occur earlier.

This fixes a potential crash in registering for multiple callbacks, and a leak when reigstering for callbacks other than EndOfEvent.